refactor(honeycrisp): inline derived lookups and delete utils.ts
- Replace headingKeys/headingValues/listKeys/listValues machinery with
  direct if-guard $derived.by expressions (4 data structures + .find()
  loops → 4 lines of if-guards each)
- Move extractTitleAndPreview into Editor.svelte and delete utils.ts
  (single consumer, no reason for a separate file)
